+++ src/main/java/io/github/mapepire_ibmi/types/SetConfigResult.java
@@ -51,7 +51,7 @@ public class SetConfigResult extends ServerResponse {
      * @param jtOpenTraceDest  The JTOpen trace data destination.
      * @param jtOpenTraceLevel The JTOpen trace level.
      */
-    public SetConfigResult(String id, boolean success, String error, int sqlRc, String sqlState, int executionTime,
+    public SetConfigResult(String id, boolean success, String error, int sqlRc, String sqlState, long executionTime,
             String traceDest, ServerTraceLevel traceLevel, String jtOpenTraceDest,
             ServerTraceLevel jtOpenTraceLevel) {
         super(id, success, error, sqlRc, sqlState, executionTime);
